+++ ptai-ee-tools-java/generic-client-lib/src/main/java/com/ptsecurity/appsec/ai/ee/utils/ci/integration/ptaiserver/v36/Reports.java
@@ -22,6 +22,9 @@ import java.util.*;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
+import static com.ptsecurity.appsec.ai.ee.ptai.server.projectmanagement.v36.IssuesFilter.*;
+import static org.apache.commons.lang3.ArrayUtils.isEmpty;
+
 @Slf4j
 @Getter
 @Setter
@@ -47,32 +50,242 @@ public class Reports {
         private final String value;
     }
 
+    /**
+     * PT AI project management API supports reports generation that
+     * allows using IssuesFilter parameters like issuesLevel,
+     * confirmationStatus, scanMode etc. But those fields are defined in
+     * Swagger API descriptions as single enums, not as arrays. That means
+     * that you can't generate report that contain only vulnerabilities of
+     * high and medium levels, you may define only one value: high or
+     * medium. But internal implementation of PT AI report generation API
+     * uses bit masks to support multiple fields definition.
+     * For example, issues levels are defined as follows:
+     *     None      = 0,
+     *     Low       = 1 << 0,
+     *     Medium    = 1 << 1,
+     *     High      = 1 << 2,
+     *     Potential = 1 << 3,
+     *     All       = Low | Medium | High | Potential
+     * That means that it is possible to define more than one field value to
+     * solve task described above. To do that I've changed Swagger API definition
+     * for IssuesFilter type: corresponding fields now defined as integers,
+     * not enums. But for user convenience we need to let him define JSONs using
+     * old-fashioned enum values. This class implements those values as both
+     * single- and multi-valued attributes and provides method to convert it
+     * to IssuesFilter
+     */
     @Getter
     @Setter
-    public static class IssuesFilterEx extends IssuesFilter {
+    public static class IssuesFilterEx {
+        @SerializedName(SERIALIZED_NAME_ISSUE_LEVEL)
+        private IssuesFilterLevel issueLevel;
+
         public static final String SERIALIZED_NAME_ISSUE_LEVELS = "issueLevels";
         @SerializedName(SERIALIZED_NAME_ISSUE_LEVELS)
         private IssuesFilterLevel[] issueLevels;
+
+        @SerializedName(SERIALIZED_NAME_CONFIRMATION_STATUS)
+        private IssuesFilterConfirmationStatus confirmationStatus;
 
         public static final String SERIALIZED_NAME_CONFIRMATION_STATUSES = "confirmationStatuses";
         @SerializedName(SERIALIZED_NAME_CONFIRMATION_STATUSES)
         private IssuesFilterConfirmationStatus[] confirmationStatuses;
 
+        @SerializedName(SERIALIZED_NAME_EXPLOITATION_CONDITION)
+        private IssuesFilterExploitationCondition exploitationCondition;
+
         public static final String SERIALIZED_NAME_EXPLOITATION_CONDITIONS = "exploitationConditions";
         @SerializedName(SERIALIZED_NAME_EXPLOITATION_CONDITIONS)
         private IssuesFilterExploitationCondition[] exploitationConditions;
+
+        @SerializedName(SERIALIZED_NAME_SUPPRESS_STATUS)
+        private IssuesFilterSuppressStatus suppressStatus;
 
         public static final String SERIALIZED_NAME_SUPPRESS_STATUSES = "suppressStatuses";
         @SerializedName(SERIALIZED_NAME_SUPPRESS_STATUSES)
         private IssuesFilterSuppressStatus[] suppressStatuses;
 
+        @SerializedName(SERIALIZED_NAME_SOURCE_TYPE)
+        private IssuesFilterSourceType sourceType;
+
         public static final String SERIALIZED_NAME_SOURCE_TYPES = "sourceTypes";
         @SerializedName(SERIALIZED_NAME_SOURCE_TYPES)
         private IssuesFilterSourceType[] sourceTypes;
 
+        @SerializedName(SERIALIZED_NAME_SCAN_MODE)
+        private IssuesFilterScanMode scanMode;
+
         public static final String SERIALIZED_NAME_SCAN_MODES = "scanModes";
         @SerializedName(SERIALIZED_NAME_SCAN_MODES)
         private IssuesFilterScanMode[] scanModes;
+
+        public static final String SERIALIZED_NAME_ACTUAL_STATUS = "actualStatus";
+        @SerializedName(SERIALIZED_NAME_ACTUAL_STATUS)
+        private IssuesFilterActualStatus actualStatus;
+
+        public static final String SERIALIZED_NAME_HIDE_SECOND_ORDER = "hideSecondOrder";
+        @SerializedName(SERIALIZED_NAME_HIDE_SECOND_ORDER)
+        private Boolean hideSecondOrder;
+
+        public static final String SERIALIZED_NAME_HIDE_SUSPECTED = "hideSuspected";
+        @SerializedName(SERIALIZED_NAME_HIDE_SUSPECTED)
+        private Boolean hideSuspected;
+
+        public static final String SERIALIZED_NAME_HIDE_POTENTIAL = "hidePotential";
+        @SerializedName(SERIALIZED_NAME_HIDE_POTENTIAL)
+        private Boolean hidePotential;
+
+        public static final String SERIALIZED_NAME_BY_FAVORITE = "byFavorite";
+        @SerializedName(SERIALIZED_NAME_BY_FAVORITE)
+        private Boolean byFavorite;
+
+        public static final String SERIALIZED_NAME_BY_BEST_PLACE_TO_FIX = "byBestPlaceToFix";
+        @SerializedName(SERIALIZED_NAME_BY_BEST_PLACE_TO_FIX)
+        private Boolean byBestPlaceToFix;
+
+        public static final String SERIALIZED_NAME_TYPES = "types";
+        @SerializedName(SERIALIZED_NAME_TYPES)
+        private List<IssuesFilterType> types = null;
+
+        public static final String SERIALIZED_NAME_PATH_INFO = "pathInfo";
+        @SerializedName(SERIALIZED_NAME_PATH_INFO)
+        private IssuesFilterPathInfo pathInfo;
+
+        public static final String SERIALIZED_NAME_PATTERN = "pattern";
+        @SerializedName(SERIALIZED_NAME_PATTERN)
+        private String pattern;
+
+        public static final String SERIALIZED_NAME_SELECT_ALL_LEVELS_SEPARATELY = "selectAllLevelsSeparately";
+        @SerializedName(SERIALIZED_NAME_SELECT_ALL_LEVELS_SEPARATELY)
+        private Boolean selectAllLevelsSeparately;
+
+        public static final String SERIALIZED_NAME_SELECT_ALL_CONFIRMATION_STATUS_SEPARATELY = "selectAllConfirmationStatusSeparately";
+        @SerializedName(SERIALIZED_NAME_SELECT_ALL_CONFIRMATION_STATUS_SEPARATELY)
+        private Boolean selectAllConfirmationStatusSeparately;
+
+        public static final String SERIALIZED_NAME_SELECT_ALL_EXPLOITATION_CONDITION_SEPARATELY = "selectAllExploitationConditionSeparately";
+        @SerializedName(SERIALIZED_NAME_SELECT_ALL_EXPLOITATION_CONDITION_SEPARATELY)
+        private Boolean selectAllExploitationConditionSeparately;
+
+        public static final String SERIALIZED_NAME_SELECT_ALL_SUPPRESS_STATUS_SEPARATELY = "selectAllSuppressStatusSeparately";
+        @SerializedName(SERIALIZED_NAME_SELECT_ALL_SUPPRESS_STATUS_SEPARATELY)
+        private Boolean selectAllSuppressStatusSeparately;
+
+        public static final String SERIALIZED_NAME_SELECT_ALL_SCAN_MODE_SEPARATELY = "selectAllScanModeSeparately";
+        @SerializedName(SERIALIZED_NAME_SELECT_ALL_SCAN_MODE_SEPARATELY)
+        private Boolean selectAllScanModeSeparately;
+
+        public static final String SERIALIZED_NAME_SELECT_ALL_ACTUAL_STATUS_SEPARATELY = "selectAllActualStatusSeparately";
+        @SerializedName(SERIALIZED_NAME_SELECT_ALL_ACTUAL_STATUS_SEPARATELY)
+        private Boolean selectAllActualStatusSeparately;
+
+        public IssuesFilter convert() {
+            IssuesFilter res = new IssuesFilter();
+            // No filters are defined - set ALL value
+            if (null == issueLevel && isEmpty(issueLevels))
+                res.setIssueLevel(IssuesFilterLevel.All.getValue());
+            else {
+                int rawValue = 0;
+                if (null != issueLevel)
+                    rawValue = issueLevel.getValue();
+                if (ArrayUtils.isNotEmpty(issueLevels)) {
+                    for (IssuesFilterLevel item : issueLevels)
+                        rawValue |= item.getValue();
+                }
+                res.setIssueLevel(rawValue);
+            }
+
+            // No exploitation conditions are defined - set ALL value
+            if (null == exploitationCondition && isEmpty(exploitationConditions))
+                res.setExploitationCondition(IssuesFilterExploitationCondition.All.getValue());
+            else {
+                int rawValue = 0;
+                if (null != exploitationCondition)
+                    rawValue = exploitationCondition.getValue();
+                if (ArrayUtils.isNotEmpty(exploitationConditions)) {
+                    for (IssuesFilterExploitationCondition item : exploitationConditions)
+                        rawValue |= item.getValue();
+                }
+                res.setExploitationCondition(rawValue);
+            }
+
+            // No scan modes are defined - set ALL value
+            if (null == scanMode && isEmpty(scanModes))
+                res.setScanMode(IssuesFilterScanMode.All.getValue());
+            else {
+                int rawValue = 0;
+                if (null != scanMode)
+                    rawValue = scanMode.getValue();
+                if (ArrayUtils.isNotEmpty(scanModes)) {
+                    for (IssuesFilterScanMode item : scanModes)
+                        rawValue |= item.getValue();
+                }
+                res.setScanMode(rawValue);
+            }
+
+            // No suppress statuses are defined - set ALL value
+            if (null == suppressStatus && isEmpty(suppressStatuses))
+                res.setSuppressStatus(IssuesFilterSuppressStatus.All.getValue());
+            else {
+                int rawValue = 0;
+                if (null != suppressStatus)
+                    rawValue = suppressStatus.getValue();
+                if (ArrayUtils.isNotEmpty(suppressStatuses)) {
+                    for (IssuesFilterSuppressStatus item : suppressStatuses)
+                        rawValue |= item.getValue();
+                }
+                res.setSuppressStatus(rawValue);
+            }
+
+            // No confirmation statuses are defined - set ALL value
+            if (null == confirmationStatus && isEmpty(confirmationStatuses))
+                res.setConfirmationStatus(IssuesFilterConfirmationStatus.All.getValue());
+            else {
+                int rawValue = 0;
+                if (null != confirmationStatus)
+                    rawValue = confirmationStatus.getValue();
+                if (ArrayUtils.isNotEmpty(confirmationStatuses)) {
+                    for (IssuesFilterConfirmationStatus item : confirmationStatuses)
+                        rawValue |= item.getValue();
+                }
+                res.setConfirmationStatus(rawValue);
+            }
+
+            // No source types are defined - set ALL value
+            if (null == sourceType && isEmpty(sourceTypes))
+                res.setSourceType(IssuesFilterSourceType.All.getValue());
+            else {
+                int rawValue = 0;
+                if (null != sourceType.getValue())
+                    rawValue = sourceType.getValue();
+                if (ArrayUtils.isNotEmpty(sourceTypes)) {
+                    for (IssuesFilterSourceType item : sourceTypes)
+                        rawValue |= item.getValue();
+                }
+                res.setSourceType(rawValue);
+            }
+
+            res.setActualStatus(null == actualStatus ? IssuesFilterActualStatus.ALL : actualStatus);
+
+            res.setHideSecondOrder(null == hideSecondOrder ? false : hideSecondOrder);
+            res.setHideSuspected(null == hideSuspected ? false : hideSuspected);
+            res.setHidePotential(null == hidePotential ? false : hidePotential);
+
+            res.setByFavorite(null == byFavorite ? false : byFavorite);
+            res.setByBestPlaceToFix(null == byBestPlaceToFix ? false : byBestPlaceToFix);
+
+            res.setTypes(types);
+            res.setPattern(pattern);
+
+            res.setSelectAllLevelsSeparately(null == selectAllLevelsSeparately ? false : selectAllLevelsSeparately);
+            res.setSelectAllConfirmationStatusSeparately(null == selectAllConfirmationStatusSeparately ? false : selectAllConfirmationStatusSeparately);
+            res.setSelectAllExploitationConditionSeparately(null == selectAllExploitationConditionSeparately ? false : selectAllExploitationConditionSeparately);
+            res.setSelectAllSuppressStatusSeparately(null == selectAllSuppressStatusSeparately ? false : selectAllSuppressStatusSeparately);
+            res.setSelectAllScanModeSeparately(null == selectAllScanModeSeparately ? false : selectAllScanModeSeparately);
+            res.setSelectAllActualStatusSeparately(null == selectAllActualStatusSeparately ? false : selectAllActualStatusSeparately);
+
+            return res;
+        }
     }
 
     @Getter
@@ -99,106 +312,6 @@ public class Reports {
         @JsonProperty
         protected IssuesFilterEx filters = null;
 
-        /**
-         * All the filters that use enum values are treated as NONE if no value
-         * is defined in JSON. But this is not convenient for user as he thinks that if
-         * no filter is defined then filtering must not be done on that field. So we need
-         * to fix all the missing filters with ALL enum
-         *
-         * @return Fixed NamedReportDefinition where all missing fields are filled with "ALL" value
-         */
-        public AbstractReport fix() {
-            if (null == filters) return this;
-
-            // No filters are defined - set ALL value
-            if (null == filters.getIssueLevel() && ArrayUtils.isEmpty(filters.getIssueLevels()))
-                filters.setIssueLevel(IssuesFilterLevel.All.getValue());
-            else {
-                int rawValue = 0;
-                if (null != filters.getIssueLevel())
-                    rawValue = filters.getIssueLevel();
-                if (ArrayUtils.isNotEmpty(filters.getIssueLevels())) {
-                    for (IssuesFilterLevel item : filters.getIssueLevels())
-                        rawValue |= item.getValue();
-                }
-                filters.setIssueLevel(rawValue);
-            }
-
-            // No exploitation conditions are defined - set ALL value
-            if (null == filters.getExploitationCondition() && ArrayUtils.isEmpty(filters.getExploitationConditions()))
-                filters.setExploitationCondition(IssuesFilterExploitationCondition.All.getValue());
-            else {
-                int rawValue = 0;
-                if (null != filters.getExploitationCondition())
-                    rawValue = filters.getExploitationCondition();
-                if (ArrayUtils.isNotEmpty(filters.getExploitationConditions())) {
-                    for (IssuesFilterExploitationCondition item : filters.getExploitationConditions())
-                        rawValue |= item.getValue();
-                }
-                filters.setExploitationCondition(rawValue);
-            }
-
-            // No scan modes are defined - set ALL value
-            if (null == filters.getScanMode() && ArrayUtils.isEmpty(filters.getScanModes()))
-                filters.setScanMode(IssuesFilterScanMode.All.getValue());
-            else {
-                int rawValue = 0;
-                if (null != filters.getScanMode())
-                    rawValue = filters.getScanMode();
-                if (ArrayUtils.isNotEmpty(filters.getScanModes())) {
-                    for (IssuesFilterScanMode item : filters.getScanModes())
-                        rawValue |= item.getValue();
-                }
-                filters.setScanMode(rawValue);
-            }
-
-            // No suppress statuses are defined - set ALL value
-            if (null == filters.getSuppressStatus() && ArrayUtils.isEmpty(filters.getSuppressStatuses()))
-                filters.setSuppressStatus(IssuesFilterSuppressStatus.All.getValue());
-            else {
-                int rawValue = 0;
-                if (null != filters.getSuppressStatus())
-                    rawValue = filters.getSuppressStatus();
-                if (ArrayUtils.isNotEmpty(filters.getSuppressStatuses())) {
-                    for (IssuesFilterSuppressStatus item : filters.getSuppressStatuses())
-                        rawValue |= item.getValue();
-                }
-                filters.setSuppressStatus(rawValue);
-            }
-
-            // No confirmation statuses are defined - set ALL value
-            if (null == filters.getConfirmationStatus() && ArrayUtils.isEmpty(filters.getConfirmationStatuses()))
-                filters.setConfirmationStatus(IssuesFilterConfirmationStatus.All.getValue());
-            else {
-                int rawValue = 0;
-                if (null != filters.getConfirmationStatus())
-                    rawValue = filters.getConfirmationStatus();
-                if (ArrayUtils.isNotEmpty(filters.getConfirmationStatuses())) {
-                    for (IssuesFilterConfirmationStatus item : filters.getConfirmationStatuses())
-                        rawValue |= item.getValue();
-                }
-                filters.setConfirmationStatus(rawValue);
-            }
-
-            // No source types are defined - set ALL value
-            if (null == filters.getSourceType() && ArrayUtils.isEmpty(filters.getSourceTypes()))
-                filters.setSourceType(IssuesFilterSourceType.All.getValue());
-            else {
-                int rawValue = 0;
-                if (null != filters.getSourceType())
-                    rawValue = filters.getSourceType();
-                if (ArrayUtils.isNotEmpty(filters.getSourceTypes())) {
-                    for (IssuesFilterSourceType item : filters.getSourceTypes())
-                        rawValue |= item.getValue();
-                }
-                filters.setSourceType(rawValue);
-            }
-
-            if (null == filters.getActualStatus())
-                filters.setActualStatus(IssuesFilterActualStatus.ALL);
-
-            return this;
-        }
     }
 
     /**
@@ -323,16 +436,6 @@ public class Reports {
     @JsonProperty
     protected List<RawData> raw = new ArrayList<>();
 
-    /**
-     * See {@link AbstractReport#fix()}
-     * @return "This" reports instance with fixed report items
-     */
-    public Reports fix() {
-        for (AbstractReport r : report) r.fix();
-        for (AbstractReport r : data) r.fix();
-        return this;
-    }
-
     public Reports check(@NonNull final Utils utils) {
         List<ImmutablePair<Locale, String>> missingTemplates = new ArrayList<>();
         // We will download all the templates for supported locales to give hint to user in case of typo
@@ -377,8 +480,8 @@ public class Reports {
         try {
             ObjectMapper mapper = new ObjectMapper();
             mapper.enable(JsonParser.Feature.ALLOW_COMMENTS);
-            IssuesFilterEx res = mapper.readValue(json, IssuesFilterEx.class);
-            return res;
+            mapper.enable(MapperFeature.ACCEPT_CASE_INSENSITIVE_ENUMS);
+            return mapper.readValue(json, IssuesFilterEx.class);
         } catch (Exception e) {
             throw ApiException.raise("JSON filter settings parse failed", e);
         }
@@ -432,8 +535,7 @@ public class Reports {
             ObjectMapper mapper = new ObjectMapper();
             mapper.enable(JsonParser.Feature.ALLOW_COMMENTS);
             mapper.enable(MapperFeature.ACCEPT_CASE_INSENSITIVE_ENUMS);
-            Reports res = mapper.readValue(json, Reports.class);
-            return res.fix();
+            return mapper.readValue(json, Reports.class);
         } catch (Exception e) {
             throw ApiException.raise(Resources.i18n_ast_result_reporting_json_message_file_parse_failed(), e);
         }