+++ ptai-ee-tools-java/generic-client-lib/src/test/java/com/ptsecurity/appsec/ai/ee/utils/ci/integration/ptaiserver/v36/ReportsTest.java
@@ -46,8 +46,6 @@ class ReportsTest {
         report.setFileName("report.ru.xml");
         reports.getReport().add(report);
 
-        reports.fix();
-
         System.out.println(new ObjectMapper().writerWithDefaultPrettyPrinter().writeValueAsString(reports));
     }