+++ ptai-ee-tools-java/ptai-jenkins-plugin/src/main/resources/com/ptsecurity/appsec/ai/ee/utils/ci/integration/plugin/jenkins/reports/Data/config.groovy
@@ -19,7 +19,9 @@ f.entry(
 f.entry(
         title: _('locale'),
         field: 'locale') {
-    f.select()
+    f.select(
+            default: descriptor.getDefaultLocale(),
+    )
 }
 
 f.advanced() {