+++ ptai-ee-tools-java/ptai-teamcity-plugin/ptai-teamcity-plugin-server/src/main/java/com/ptsecurity/appsec/ai/ee/utils/ci/integration/plugin/teamcity/service/AstSettingsService.java
@@ -342,7 +342,7 @@ public class AstSettingsService {
             }
             // Check reporting settings
             Reports reports = bean.convert();
-            reports.validate().fix().check(utils);
+            reports.validate().check(utils);
         } catch (ApiException e) {
             log.warn(e.getDetailedMessage(), e);
             results.add(e);